requirements.txt
anticaptchaofficial==1.0.56
mysql-connector-python==8.3.0
ruamel.std.zipfile==0.1.0
schedule==1.2.1
selenium==4.8.3


